+++ src/Ydb.Sdk/src/Auth/ICredentialsProvider.cs
@@ -4,7 +4,13 @@ namespace Ydb.Sdk.Auth;
 
 public interface ICredentialsProvider
 {
+    [Obsolete("For removal in 1.*")]
     string? GetAuthInfo();
+
+    ValueTask<string?> GetAuthInfoAsync()
+    {
+        return ValueTask.FromResult(GetAuthInfo());
+    }
 
     Task ProvideAuthClient(AuthClient authClient)
     {